backend/requirements.txt

Django==5.0.2
djangorestframework==3.14.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
pytest==7.4.4
pytest-django==4.7.0
